improve disposing / removing objects improve bounds builder improve shadow map with dynamic volume improve stereo camera / multiview fix panorma ibl improve gltf loader

+++ src/Engine/XrEngine.Core/Shaders/PbrV2/spbrdf_cs.glsl
@@ -1,5 +1,5 @@
 
-precision mediump image2D;
+precision highp image2D;
 
 // Physically Based Rendering
 // Copyright (c) 2017-2018 Michał Siejak
